ryacast prototyp v1.3

+++ utils2.c
@@ -6,13 +6,13 @@
 /*   By: haouky <haouky@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/10 12:13:37 by haouky            #+#    #+#             */
-/*   Updated: 2024/12/12 11:49:06 by haouky           ###   ########.fr       */
+/*   Updated: 2024/12/15 09:57:23 by haouky           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "cub3D.h"
 
-void map_max_sz(char **map,double *size)
+void map_max_sz(char **map,int *size)
 {
     int y;
     int tp;
